sample-fastapi-graphql/pyproject.toml
[tool.poetry]
name = "fastapi-gql"
version = "0.0.1"
description = "Sample GraphQL Application with FastAPI + Strawberry + EdgeDB, OPA(Open Policy Agent)"
authors = ["Haze Lee <hazelee@realignist.me>"]

[tool.black]
line-length = 80
skip_string_normalization = false

[tool.zimports]
black-line-length = 80
keep-unused-type-checking = false
multi-imports = true

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
anyio = "^3.6.1"
pydantic = {extras = ["email"], version = "^1.9.1"}
edgedb = "^0.24.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.124.0"}
# TODO(Hazealign): uvicorn 0.18 [standard] requires to build with rust toolchain.
uvicorn = { extras = ["standard"], version = "~0.17.6"}
pendulum = "^2.1.2"
click = "^8.1.3"
python-dotenv = "^0.20.0"
PyJWT = "^2.4.0"
argon2-cffi = "^21.3.0"

[tool.poetry.dev-dependencies]
mypy = "^0.971"
pytest = "^7.1.2"
pytest-anyio = "^0.0.0"
pytest-cov = "^3.0.0"
flake8 = "^5.0.4"
flake8-import-order = "^0.18.1"
pre-commit = "^2.20.0"
black = "^22.6.0"
zimports = "^0.6.0"
yesqa = "^1.4.0"
types-PyJWT = "^1.7.1"
isort = "^5.10.1"

[tool.poetry.scripts]
server = "fastapi_gql.server:cli_main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
